Updated dungeoncrawl.capx
I've recreated the Canal Level in dungeoncrawl.capx and I've connected
it to Room 2 (the door that was originally closed closest to tentacle
monster). I got the hero to move around inside my level, however, I'm
currently trying to figure out why he's showing up behind the door and
not in front of it.